Changed write_to_syslog, send_event return types to Status
With a Result<(), Status>, the compiler warns if you don't use the value.
But if the write_to_syslog or send_event doesn't succeed, what are
you supposed to do with the result? The paranoid can use Status.as_result(),
while the rest of us can just silently drop the value.