Changed Message.set_msgid to make it safe
Before said changes, a user of this crate could change the message ID
of a Message from that of a command packet to that of a telemetry packet
(or vice versa) in safe code. Said change could lead to future calls
into this crate doing undesirable things, though.

As such, we now check that set_msgid() doesn't change the message type
of the Message before going ahead with the operation.