CMake: set '--jobs 1' in cargo invocations
Previously, each of the invocations of Cargo would try running with up
to [number of cores on machine] parallel tasks. This can cause high load
if you're building multiple apps simulataneously.

The *best* solution would be to give each Cargo instantiation access to
Make's jobserver so as to coordinate concurrency control while still
making use of internal parallelism as appropiate...
but CMake doesn't really make that accessible in the general case. So
instead, we constrain each Cargo run to only run one job at a time. Not
ideal, but it's what we have available.