Moved cfs-sys functionality into n2o4 crate
Merging what is currently the separate cfs-sys crate into the n2o4 crate
is being done for three purposes:

1) At this point, I don't think it likely that anyone would want to use
cfs-sys, except wrapped by n2o4. Certainly, the way n2o4 is
distributed (or not) means it is always tagging along cfs-sys.

2) In the future, we'd like to make n2o4 work with multiple versions of
cFE and OSAL (issue BlackCAT-CubeSat#1). My best ideas for how to do
that would require a merger of the cfs-sys and n2o4 crates anyway.

3) It reduces the file-hierarchy depth and breadth slightly--
fewer places for stuff to be in, basically.